+++ driving_log_replayer/scripts/eagleye_evaluator_node.py
@@ -14,24 +14,17 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
 import os
 from typing import TYPE_CHECKING
 
-from autoware_adapi_v1_msgs.srv import InitializeLocalization
 from diagnostic_msgs.msg import DiagnosticArray
 from diagnostic_msgs.msg import DiagnosticStatus
 import rclpy
-from rclpy.callback_groups import MutuallyExclusiveCallbackGroup
-from rclpy.clock import Clock
-from rclpy.clock import ClockType
-from rclpy.executors import MultiThreadedExecutor
 from rclpy.node import Node
-from rclpy.time import Time
 from tier4_localization_msgs.srv import PoseWithCovarianceStamped
 import yaml
 
-from driving_log_replayer.node_common import set_initial_pose
+from driving_log_replayer.evaluator import DLREvaluator
 from driving_log_replayer.result import ResultBase
 from driving_log_replayer.result import ResultWriter
 
@@ -78,56 +71,21 @@ class EagleyeResult(ResultBase):
             self.update()
 
 
-class EagleyeEvaluator(Node):
-    def __init__(self):
-        super().__init__("eagleye_evaluator")
-        self.declare_parameter("scenario_path", "")
-        self.declare_parameter("result_json_path", "")
-
-        self.__timer_group = MutuallyExclusiveCallbackGroup()
-
-        scenario_path = os.path.expandvars(
-            self.get_parameter("scenario_path").get_parameter_value().string_value
-        )
-        self.__scenario_yaml_obj = None
-        with open(scenario_path) as scenario_file:
-            self.__scenario_yaml_obj = yaml.safe_load(scenario_file)
-        self.__result_json_path = os.path.expandvars(
-            self.get_parameter("result_json_path").get_parameter_value().string_value
-        )
+class EagleyeEvaluator(DLREvaluator):
+    def __init__(self, name: str):
+        super().__init__(name)
 
         self.__result = EagleyeResult()
-
         self.__result_writer = ResultWriter(self.__result_json_path, self.get_clock(), {})
 
-        self.__initial_pose = set_initial_pose(
-            self.__scenario_yaml_obj["Evaluation"]["InitialPose"]
-        )
-        self.__initial_pose_running = False
-        self.__initial_pose_success = False
+        self._scenario_yaml_obj = None
+        with open(self._scenario_path) as scenario_file:
+            self._scenario_yaml_obj = yaml.safe_load(scenario_file)
 
-        self.__current_time = Time().to_msg()
-        self.__prev_time = Time().to_msg()
-        self.__counter = 0
-
-        # service client
-        self.__initial_pose_client = self.create_client(
-            InitializeLocalization, "/api/localization/initialize"
+        self._initial_pose = DLREvaluator.set_initial_pose(
+            self._scenario_yaml_obj["Evaluation"]["InitialPose"]
         )
-        self.__map_fit_client = self.create_client(
-            PoseWithCovarianceStamped, "/map/map_height_fitter/service"
-        )
-        while not self.__initial_pose_client.wait_for_service(timeout_sec=1.0):
-            self.get_logger().warning("initial pose service not available, waiting again...")
-        while not self.__map_fit_client.wait_for_service(timeout_sec=1.0):
-            self.get_logger().warning("map height fitter service not available, waiting again...")
-
-        self.__timer = self.create_timer(
-            1.0,
-            self.timer_cb,
-            callback_group=self.__timer_group,
-            clock=Clock(clock_type=ClockType.SYSTEM_TIME),
-        )  # wall timer
+        self.start_initial_pose()
 
         self.__sub_diagnostics = self.create_subscription(
             DiagnosticArray,
@@ -141,72 +99,33 @@ class EagleyeEvaluator(Node):
         self.__result_writer.write(self.__result)
 
     def timer_cb(self) -> None:
-        self.__current_time = self.get_clock().now().to_msg()
+        self._current_time = self.get_clock().now().to_msg()
         # self.get_logger().error(f"time: {self.__current_time.sec}.{self.__current_time.nanosec}")
-        if self.__current_time.sec > 0:
+        if self._current_time.sec > 0:
             if (
-                self.__initial_pose is not None
-                and not self.__initial_pose_success
-                and not self.__initial_pose_running
+                self._initial_pose is not None
+                and not self._initial_pose_success
+                and not self._initial_pose_running
             ):
                 self.get_logger().info(
-                    f"call initial_pose time: {self.__current_time.sec}.{self.__current_time.nanosec}"
+                    f"call initial_pose time: {self._current_time.sec}.{self._current_time.nanosec}"
                 )
-                self.__initial_pose_running = True
-                self.__initial_pose.header.stamp = self.__current_time
+                self._initial_pose_running = True
+                self._initial_pose.header.stamp = self.__current_time
                 future_map_fit = self.__map_fit_client.call_async(
                     PoseWithCovarianceStamped.Request(pose_with_covariance=self.__initial_pose)
                 )
                 future_map_fit.add_done_callback(self.map_fit_cb)
-            if self.__current_time == self.__prev_time:
-                self.__counter += 1
+            if self._current_time == self.__prev_time:
+                self._clock_stop_counter += 1
             else:
-                self.__counter = 0
-            self.__prev_time = self.__current_time
-            if self.__counter >= 5:
-                self.__result_writer.close()
+                self._clock_stop_counter = 0
+            self._prev_time = self._current_time
+            if self._clock_stop_counter >= 5:
+                self._result_writer.close()
                 rclpy.shutdown()
-
-    def map_fit_cb(self, future):
-        result = future.result()
-        if result is not None:
-            if result.success:
-                future_init_pose = self.__initial_pose_client.call_async(
-                    InitializeLocalization.Request(pose=[result.pose_with_covariance])
-                )
-                future_init_pose.add_done_callback(self.initial_pose_cb)
-            else:
-                # free self.__initial_pose_running when the service call fails
-                self.__initial_pose_running = False
-                self.get_logger().warn("map_height_height service result is fail")
-        else:
-            # free self.__initial_pose_running when the service call fails
-            self.__initial_pose_running = False
-            self.get_logger().error(f"Exception for service: {future.exception()}")
-
-    def initial_pose_cb(self, future):
-        result = future.result()
-        if result is not None:
-            res_status: ResponseStatus = result.status
-            self.__initial_pose_success = res_status.success
-            self.get_logger().info(
-                f"initial_pose_success: {self.__initial_pose_success}"
-            )  # debug msg
-        else:
-            self.get_logger().error(f"Exception for service: {future.exception()}")
-        # free self.__initial_pose_running
-        self.__initial_pose_running = False
-
-
-def main(args=None):
-    rclpy.init(args=args)
-    executor = MultiThreadedExecutor()
-    localization_evaluator = EagleyeEvaluator()
-    executor.add_node(localization_evaluator)
-    executor.spin()
-    localization_evaluator.destroy_node()
-    rclpy.shutdown()
 
 
 if __name__ == "__main__":
-    main()
+    evaluator = EagleyeEvaluator("eagleye_evaluator")
+    evaluator.run()
